# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Verana Indexer API",
    "description": "Comprehensive API for managing decentralized identifiers, trust registries, credential schemas, and permissions in the Verana ecosystem.\n\nIndexer endpoints (under /verana/indexer/v1) expose the current indexed block height and a changes feed that can be queried by block_height. Other module endpoints may accept the At-Block-Height header to query historical state; when omitted, they return the latest indexed state.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    },
    {
      "url": "https://idx.devnet.verana.network",
      "description": "Devnet API server"
    },
    {
      "url": "https://idx.testnet.verana.network",
      "description": "Testnet API server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/verana/indexer/v1/block-height": {
      "get": {
        "tags": [
          "Indexer"
        ],
        "summary": "Get indexed block height",
        "description": "Return the latest block height that has been fully processed by the indexer.",
        "operationId": "getBlockHeight",
        "responses": {
          "200": {
            "description": "Current indexed block height",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Response type identifier",
                      "example": "block-processed"
                    },
                    "height": {
                      "type": "integer",
                      "minimum": 0,
                      "description": "The latest block height that has been fully processed by the indexer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "The timestamp when the checkpoint was last updated (ISO 8601 format without milliseconds)"
                    }
                  },
                    "required": ["type", "height", "timestamp"]
                },
                "example": {
                  "type": "block-processed",
                  "height": 1794800,
                  "timestamp": "2025-12-17T04:42:55Z"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/verana/indexer/v1/status": {
      "get": {
        "tags": [
          "Indexer"
        ],
        "summary": "Get indexer operational status",
        "description": "Return the current operational status of the indexer, including whether it is running, whether crawling is active, and any error information if crawling or the indexer has stopped. This endpoint is useful for monitoring and debugging indexer issues.",
        "operationId": "getIndexerStatus",
        "responses": {
          "200": {
            "description": "Indexer status information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isRunning": {
                      "type": "boolean",
                      "description": "Whether the indexer service is running. If false, all APIs will return 503.",
                      "example": true
                    },
                    "isCrawling": {
                      "type": "boolean",
                      "description": "Whether the indexer is actively crawling new blocks and transactions. If false, APIs remain available but no new data is being indexed.",
                      "example": false
                    },
                    "stoppedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp when the indexer or crawling stopped (if applicable)",
                      "example": "2024-01-15T10:30:00Z"
                    },
                    "stoppedReason": {
                      "type": "string",
                      "description": "Human-readable reason why crawling or the indexer stopped (if applicable)",
                      "example": "Crawling stopped due to: Unknown Verana message types: /verana.perm.v1.MsgCreateRootPermission. Indexer APIs remain available."
                    },
                    "lastError": {
                      "type": "object",
                      "description": "Details about the last error that caused crawling or indexer to stop (if applicable)",
                      "properties": {
                        "message": {
                          "type": "string",
                          "description": "Error message",
                          "example": "Unknown Verana message types: /verana.perm.v1.MsgCreateRootPermission"
                        },
                        "stack": {
                          "type": "string",
                          "description": "Error stack trace (if available)",
                          "example": "Error: Unknown Verana message types...\n    at ..."
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time",
                          "description": "When the error occurred",
                          "example": "2024-01-15T10:30:00Z"
                        },
                        "service": {
                          "type": "string",
                          "description": "Service name where the error occurred",
                          "example": "crawl-tx"
                        }
                      }
                    }
                  },
                  "required": [
                    "isRunning",
                    "isCrawling"
                  ]
                },
                "example": {
                  "isRunning": true,
                  "isCrawling": false,
                  "stoppedAt": "2024-01-15T10:30:00Z",
                  "stoppedReason": "Crawling stopped due to: Unknown Verana message types: /verana.perm.v1.MsgCreateRootPermission. Indexer APIs remain available.",
                  "lastError": {
                    "message": "Unknown Verana message types: /verana.perm.v1.MsgCreateRootPermission",
                    "timestamp": "2024-01-15T10:30:00Z",
                    "service": "crawl-tx"
                  }
                }
              }
            },
            "headers": {
              "X-Crawling-Status": {
                "description": "Current crawling status: 'active' or 'stopped'",
                "schema": {
                  "type": "string",
                  "enum": [
                    "active",
                    "stopped"
                  ],
                  "example": "stopped"
                }
              },
              "X-Indexer-Status": {
                "description": "Current indexer status: 'running' or 'stopped'",
                "schema": {
                  "type": "string",
                  "enum": [
                    "running",
                    "stopped"
                  ],
                  "example": "running"
                }
              },
              "X-Crawling-Reason": {
                "description": "Reason why crawling stopped (only present when crawling is stopped)",
                "schema": {
                  "type": "string",
                  "example": "Crawling stopped due to: Unknown Verana message types..."
                }
              },
              "X-Crawling-Error": {
                "description": "Error message that caused crawling to stop (only present when crawling is stopped)",
                "schema": {
                  "type": "string",
                  "example": "Unknown Verana message types: /verana.perm.v1.MsgCreateRootPermission"
                }
              },
              "X-Crawling-Stopped-At": {
                "description": "ISO 8601 timestamp when crawling stopped (only present when crawling is stopped)",
                "schema": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/verana/indexer/v1/changes/{block_height}": {
      "get": {
        "400": {
          "$ref": "#/components/responses/BadRequest"
        },
        "500": {
          "$ref": "#/components/responses/ServerError"
        },
        "tags": [
          "Indexer"
        ],
        "summary": "List entity changes at a block height",
        "description": "Return all indexed entity changes for the requested block height.",
        "operationId": "listChanges",
        "parameters": [
          {
            "name": "block_height",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "The block height to query changes for. Must be a positive integer not exceeding the current indexed block height."
          }
        ],
        "responses": {
          "200": {
            "description": "Entity changes recorded at the requested block height",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "block_height": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "next_change_at": {
                      "type": "integer",
                      "minimum": 0,
                      "nullable": true,
                      "description": "The next block height greater than the requested block_height that has any indexed change events. Null when there is no later change."
                    },
                    "activity": {
                      "type": "array",
                      "description": "List of activity items at this block height (same format as history endpoints)",
                      "items": {
                        "$ref": "#/components/schemas/ActivityItem"
                      }
                    }
                  }
                },
                "example": {
                  "block_height": 63120,
                  "next_change_at": 63121,
                  "activity": [
                    {
                      "timestamp": "2025-10-06T08:46:20.686Z",
                      "block_height": 63120,
                      "entity_type": "DID",
                      "entity_id": "did:Test:123",
                      "msg": "AddDID",
                      "changes": {
                        "did": "did:Test:123",
                        "exp": "2026-10-06T08:46:20.686Z",
                        "years": 1,
                        "created": "2025-10-06T08:46:20.686Z",
                        "deposit": 5000000,
                        "modified": "2025-10-06T08:46:20.686Z",
                        "controller": "verana1evvrzxw9yg5staqdvumd6fupy3jhaxfflla7st",
                        "event_type": "/verana.dd.v1.MsgAddDID",
                        "is_deleted": false
                      },
                      "account": "verana1evvrzxw9yg5staqdvumd6fupy3jhaxfflla7st"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/verana/dd/v1/get/{did}": {
      "get": {
        "tags": [
          "DID"
        ],
        "summary": "Get a DID",
        "description": "Retrieve a specific Decentralized Identifier document by its DID",
        "operationId": "getDID",
        "parameters": [
          {
            "name": "did",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The Decentralized Identifier to retrieve"
          },
          {
            "$ref": "#/components/parameters/At-Block-Height"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with DID document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DIDResponse"
                },
                "example": {
                  "did": {
                    "did": "did:Test:123",
                    "controller": "verana1evvrzxw9yg5staqdvumd6fupy3jhaxfflla7st",
                    "deposit": 5000000,
                    "exp": "2026-10-06T08:46:20.686Z",
                    "created": "2025-10-06T08:46:20.686Z",
                    "modified": "2025-10-06T08:46:20.686Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "DID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "DID not found: did:Test:123",
                  "code": 404
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/verana/dd/v1/params": {
      "get": {
        "tags": [
          "DID"
        ],
        "summary": "Get DID parameters",
        "description": "Retrieve current system parameters for DID directory",
        "operationId": "getDIDParams",
        "parameters": [
          {
            "$ref": "#/components/parameters/At-Block-Height"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with DID parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DIDParamsResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/verana/tr/v1/get/{id}": {
      "get": {
        "tags": [
          "Trust Registry"
        ],
        "summary": "Get Trust Registry",
        "description": "Retrieve a specific Trust Registry by its ID",
        "operationId": "getTrustRegistry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "The Trust Registry ID to retrieve"
          },
          {
            "name": "active_gf_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Filter only active governance frameworks"
          },
          {
            "name": "preferred_language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Preferred language for document content"
          },
          {
            "$ref": "#/components/parameters/At-Block-Height"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with Trust Registry details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustRegistryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Trust Registry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Trust Registry not found: 1",
                  "code": 404
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/verana/cs/v1/get/{id}": {
      "get": {
        "tags": [
          "Credential Schema"
        ],
        "summary": "Get Credential Schema",
        "description": "Retrieve a specific Credential Schema by its ID",
        "operationId": "getCredentialSchema",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "The Credential Schema ID to retrieve"
          },
          {
            "$ref": "#/components/parameters/At-Block-Height"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with Credential Schema details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Credential Schema not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Credential Schema not found: 3",
                  "code": 404
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
...
    "parameters": {
      "At-Block-Height": {
        "name": "At-Block-Height",
        "in": "header",
        "required": false,
        "description": "Header to query historical state at a specific block height. When provided, returns the state as it was when the specified block finished processing. When omitted, returns the latest indexed state. Must be a positive integer not exceeding the current indexed block height.",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      },
...
      "DIDResponse": {
        "type": "object",
        "properties": {
          "did": {
            "$ref": "#/components/schemas/DID"
          }
        }
      },
...
      "TrustRegistryResponse": {
        "type": "object",
        "properties": {
          "trust_registry": {
            "$ref": "#/components/schemas/TrustRegistry"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid parameters provided",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Invalid parameter: response_max_size must be between 1 and 1024",
              "code": 400
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error occurred",
              "code": 500
            }
          }
        }
      }
    }
  },
...
}